+++ engine/previews/citizens_advice_components/disclosure_preview.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module CitizensAdviceComponents
+  class DisclosurePreview < ViewComponent::Preview
+    def example; end
+
+    def without_open_summary; end
+  end
+end